@@ -1,7 +1,8 @@
 package com.example.application.data.entity;
 
-import javax.persistence.*;
-import javax.validation.constraints.Email;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+
 import java.time.LocalDate;
 
 @Entity
