@@ -1,17 +1,17 @@
 package com.example.application.data.entity;
 
+import jakarta.annotation.Nullable;
+import jakarta.persistence.Entity;
+import jakarta.persistence.OneToMany;
+import jakarta.validation.constraints.NotBlank;
+
 import java.util.LinkedList;
 import java.util.List;
-import javax.annotation.Nullable;
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
-import javax.validation.constraints.NotBlank;
 
 @Entity
 public class Company extends AbstractEntity {
     @NotBlank
     private String name;
-
     @OneToMany(mappedBy = "company")
     @Nullable
     private List<Contact> employees = new LinkedList<>();
