@@ -12,8 +12,8 @@ import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.value.ValueChangeMode;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
+import jakarta.annotation.security.RolesAllowed;
 
-import javax.annotation.security.RolesAllowed;
 
 @PageTitle("Configuration")
 @Route(value = "config", layout= MainLayout.class)
