@@ -11,8 +11,8 @@ import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.value.ValueChangeMode;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
+import jakarta.annotation.security.PermitAll;
 
-import javax.annotation.security.PermitAll;
 
 
 @PageTitle("Info | by DBUSS GmbH")
